frontend: home: Move cluster inventory helpers
