+++ frontend/src/components/App/Home/ClusterTable.tsx
@@ -34,7 +34,7 @@ import { loadTableSettings, storeTableSettings } from '../../../helpers/tableSet
 import { formatClusterPathParam } from '../../../lib/cluster';
 import { useClustersConf, useClustersVersion } from '../../../lib/k8s';
 import { ApiError } from '../../../lib/k8s/api/v2/ApiError';
-import { Cluster, KubeCondition } from '../../../lib/k8s/cluster';
+import { Cluster } from '../../../lib/k8s/cluster';
 import { createRouteURL } from '../../../lib/router/createRouteURL';
 import { getClusterPrefixedPath } from '../../../lib/util';
 import { useTypedSelector } from '../../../redux/hooks';
@@ -45,102 +45,16 @@ import { LightTooltip } from '../../common/Tooltip';
 import { useLocalStorageState } from '../../globalSearch/useLocalStorageState';
 import ClusterBadge from '../../Sidebar/ClusterBadge';
 import ClusterContextMenu from './ClusterContextMenu';
-import { canSelectCluster, getClusterStatus, getClusterStatusLabel } from './clusterStatus';
+import {
+  getClusterStatusAccessor,
+  getClusterStatusInfo,
+  getConditionTooltip,
+  isClusterInventoryCluster,
+  STATUS_VARIANTS,
+} from './ClusterInventory';
+import { canSelectCluster } from './clusterStatus';
 import { MULTI_HOME_ENABLED } from './config';
 import { getCustomClusterNames } from './customClusterNames';
-
-const CLUSTER_INVENTORY_SOURCE = 'cluster_inventory';
-const CONTROL_PLANE_HEALTHY_CONDITION = 'ControlPlaneHealthy';
-
-type ClusterInventoryCondition = Pick<
-  KubeCondition,
-  'type' | 'status' | 'reason' | 'message' | 'lastTransitionTime'
->;
-
-type ClusterStatusKind = 'active' | 'error' | 'unknown';
-
-interface ClusterStatusInfo {
-  kind: ClusterStatusKind;
-  text: string;
-  condition: ClusterInventoryCondition | null;
-}
-
-const STATUS_VARIANTS: Record<
-  ClusterStatusKind,
-  { icon: string; colorKey: 'success' | 'error' | 'unknown'; coloredText: boolean }
-> = {
-  active: { icon: 'mdi:cloud-check-variant', colorKey: 'success', coloredText: true },
-  error: { icon: 'mdi:cloud-off', colorKey: 'error', coloredText: true },
-  unknown: { icon: 'mdi:cloud-question', colorKey: 'unknown', coloredText: false },
-};
-
-function getControlPlaneHealthyCondition(cluster: Cluster): ClusterInventoryCondition | null {
-  if (cluster?.meta_data?.source !== CLUSTER_INVENTORY_SOURCE) {
-    return null;
-  }
-
-  const conditions = cluster?.meta_data?.clusterInventory?.conditions;
-  if (!Array.isArray(conditions)) {
-    return null;
-  }
-
-  return (
-    conditions.find(
-      (condition: ClusterInventoryCondition) => condition.type === CONTROL_PLANE_HEALTHY_CONDITION
-    ) ?? null
-  );
-}
-
-function getConditionTooltip(condition: ClusterInventoryCondition): string {
-  return [condition.reason, condition.message, condition.lastTransitionTime]
-    .filter(Boolean)
-    .join('\n');
-}
-
-function getClusterStatusInfo(
-  cluster: Cluster,
-  error: ApiError | null | undefined,
-  t: (key: string) => string
-): ClusterStatusInfo {
-  const condition = getControlPlaneHealthyCondition(cluster);
-
-  if (condition?.status === 'False') {
-    return { kind: 'error', text: t('translation|Control plane unhealthy'), condition };
-  }
-
-  const status = getClusterStatus(error);
-  if (status === 'auth-error' || status === 'permission-error' || status === 'unavailable') {
-    return { kind: 'error', text: getClusterStatusLabel(t, error), condition };
-  }
-
-  if (condition?.status === 'Unknown' || status === 'loading') {
-    return { kind: 'unknown', text: '⋯', condition };
-  }
-
-  return { kind: 'active', text: getClusterStatusLabel(t, error), condition };
-}
-
-function getClusterStatusAccessor(
-  cluster: Cluster,
-  error: ApiError | null | undefined,
-  t: (key: string) => string
-): string | undefined {
-  const condition = getControlPlaneHealthyCondition(cluster);
-  if (condition?.status === 'False') {
-    return t('translation|Control plane unhealthy');
-  }
-
-  const status = getClusterStatus(error);
-  if (status === 'auth-error' || status === 'permission-error' || status === 'unavailable') {
-    return getClusterStatusLabel(t, error);
-  }
-
-  if (condition?.status === 'Unknown') {
-    return t('translation|Unknown');
-  }
-
-  return getClusterStatusLabel(t, error);
-}
 
 /**
  * ClusterStatus component displays the status of a cluster.
@@ -286,7 +200,7 @@ export default function ClusterTable({
       return t('translation|Plugin');
     } else if (cluster?.meta_data?.source === 'incluster') {
       return t('translation|In-cluster');
-    } else if (cluster?.meta_data?.source === CLUSTER_INVENTORY_SOURCE) {
+    } else if (isClusterInventoryCluster(cluster)) {
       return t('translation|Cluster Inventory');
     }
     return t('translation|Unknown');
