+++ frontend/src/i18n/locales/ar/translation.json
@@ -59,12 +59,15 @@
   "Settings": "الإعدادات",
   "Delete": "حذف",
   "Delete Cluster": "حذف الكتلة",
+  "Control plane unhealthy": "",
+  "Unknown": "غير معروف",
   "Active": "نشط",
   "Authentication required": "",
   "Insufficient permissions": "",
   "Unavailable": "",
   "Plugin": "الإضافة",
   "In-cluster": "داخل الكتلة",
+  "Cluster Inventory": "",
   "View Clusters": "عرض الكتل",
   "Loading...": "جارٍ التحميل...",
   "No clusters found": "لم يتم العثور على أي كتل",
@@ -94,7 +97,6 @@
   "Show notifications": "عرض الإشعارات",
   "You have unread notifications": "لديك إشعارات غير مقروءة",
   "View all notifications": "عرض جميع الإشعارات",
-  "Unknown": "غير معروف",
   "Are you sure you want to delete this plugin?": "هل أنت متأكد أنك تريد حذف هذه الإضافة؟",
   "Plugins": "الإضافات",
   "Multiple versions": "إصدارات متعددة",
