+++ frontend/src/i18n/locales/de/translation.json
@@ -59,11 +59,12 @@
   "Settings": "Einstellungen",
   "Delete": "Löschen",
   "Delete Cluster": "",
+  "Control plane unhealthy": "",
+  "Unknown": "",
   "Active": "Aktiv",
   "Authentication required": "",
   "Insufficient permissions": "",
   "Unavailable": "",
-  "Control plane unhealthy": "",
   "Plugin": "",
   "In-cluster": "",
   "Cluster Inventory": "",
@@ -96,7 +97,6 @@
   "Show notifications": "Nachrichten",
   "You have unread notifications": "Sie haben ungelesene Nachrichten",
   "View all notifications": "Alle Benachrichtigungen anzeigen",
-  "Unknown": "",
   "Are you sure you want to delete this plugin?": "",
   "Plugins": "Plugins",
   "Multiple versions": "",
