+++ frontend/src/i18n/locales/en/translation.json
@@ -59,11 +59,12 @@
   "Settings": "Settings",
   "Delete": "Delete",
   "Delete Cluster": "Delete Cluster",
+  "Control plane unhealthy": "Control plane unhealthy",
+  "Unknown": "Unknown",
   "Active": "Active",
   "Authentication required": "Authentication required",
   "Insufficient permissions": "Insufficient permissions",
   "Unavailable": "Unavailable",
-  "Control plane unhealthy": "Control plane unhealthy",
   "Plugin": "Plugin",
   "In-cluster": "In-cluster",
   "Cluster Inventory": "Cluster Inventory",
@@ -96,7 +97,6 @@
   "Show notifications": "Show notifications",
   "You have unread notifications": "You have unread notifications",
   "View all notifications": "View all notifications",
-  "Unknown": "Unknown",
   "Are you sure you want to delete this plugin?": "Are you sure you want to delete this plugin?",
   "Plugins": "Plugins",
   "Multiple versions": "Multiple versions",
