+++ frontend/src/i18n/locales/es/translation.json
@@ -59,11 +59,12 @@
   "Settings": "Definiciones",
   "Delete": "Borrar",
   "Delete Cluster": "Borrar cluster",
+  "Control plane unhealthy": "",
+  "Unknown": "Desconocido",
   "Active": "Activo",
   "Authentication required": "",
   "Insufficient permissions": "",
   "Unavailable": "",
-  "Control plane unhealthy": "",
   "Plugin": "Plugin",
   "In-cluster": "En-cluster",
   "Cluster Inventory": "",
@@ -96,7 +97,6 @@
   "Show notifications": "Mostrar notificaciones",
   "You have unread notifications": "Tienes notificaciones sin leer",
   "View all notifications": "Ver todas las notificaciones",
-  "Unknown": "Desconocido",
   "Are you sure you want to delete this plugin?": "¿Está seguro de que desea borrar este plugin?",
   "Plugins": "Plugins",
   "Multiple versions": "",
