+++ frontend/src/i18n/locales/fr/translation.json
@@ -59,11 +59,12 @@
   "Settings": "Paramètres",
   "Delete": "Supprimer",
   "Delete Cluster": "Supprimer le cluster",
+  "Control plane unhealthy": "",
+  "Unknown": "Inconnu",
   "Active": "Actif",
   "Authentication required": "",
   "Insufficient permissions": "",
   "Unavailable": "",
-  "Control plane unhealthy": "",
   "Plugin": "Plugin",
   "In-cluster": "Dans le cluster",
   "Cluster Inventory": "",
@@ -96,7 +97,6 @@
   "Show notifications": "Afficher les notifications",
   "You have unread notifications": "Vous avez des notifications non lues",
   "View all notifications": "Voir toutes les notifications",
-  "Unknown": "Inconnu",
   "Are you sure you want to delete this plugin?": "Êtes-vous sûr de vouloir supprimer ce plugin ?",
   "Plugins": "Plugins",
   "Multiple versions": "Versions multiples",
