+++ frontend/src/i18n/locales/he/translation.json
@@ -59,12 +59,15 @@
   "Settings": "Settings",
   "Delete": "Delete",
   "Delete Cluster": "Delete Cluster",
+  "Control plane unhealthy": "",
+  "Unknown": "Unknown",
   "Active": "Active",
   "Authentication required": "",
   "Insufficient permissions": "",
   "Unavailable": "",
   "Plugin": "Plugin",
   "In-cluster": "In-cluster",
+  "Cluster Inventory": "",
   "View Clusters": "View Clusters",
   "Loading...": "Loading...",
   "No clusters found": "No clusters found",
@@ -94,7 +97,6 @@
   "Show notifications": "Show notifications",
   "You have unread notifications": "You have unread notifications",
   "View all notifications": "View all notifications",
-  "Unknown": "Unknown",
   "Are you sure you want to delete this plugin?": "Are you sure you want to delete this plugin?",
   "Plugins": "Plugins",
   "Multiple versions": "Multiple versions",
