+++ frontend/src/i18n/locales/hi/translation.json
@@ -59,11 +59,12 @@
   "Settings": "",
   "Delete": "",
   "Delete Cluster": "",
+  "Control plane unhealthy": "",
+  "Unknown": "",
   "Active": "",
   "Authentication required": "",
   "Insufficient permissions": "",
   "Unavailable": "",
-  "Control plane unhealthy": "",
   "Plugin": "",
   "In-cluster": "",
   "Cluster Inventory": "",
@@ -96,7 +97,6 @@
   "Show notifications": "",
   "You have unread notifications": "",
   "View all notifications": "",
-  "Unknown": "",
   "Are you sure you want to delete this plugin?": "",
   "Plugins": "",
   "Multiple versions": "",
