+++ frontend/src/i18n/locales/it/translation.json
@@ -59,11 +59,12 @@
   "Settings": "Impostazioni",
   "Delete": "Elimina",
   "Delete Cluster": "Elimina Cluster",
+  "Control plane unhealthy": "",
+  "Unknown": "Sconosciuto",
   "Active": "Attivo",
   "Authentication required": "",
   "Insufficient permissions": "",
   "Unavailable": "",
-  "Control plane unhealthy": "",
   "Plugin": "Estensione",
   "In-cluster": "In-cluster",
   "Cluster Inventory": "",
@@ -96,7 +97,6 @@
   "Show notifications": "Mostra notifiche",
   "You have unread notifications": "Hai notifiche non lette",
   "View all notifications": "Visualizza tutte le notifiche",
-  "Unknown": "Sconosciuto",
   "Are you sure you want to delete this plugin?": "Sei sicuro di voler eliminare questa estensione?",
   "Plugins": "Estensioni",
   "Multiple versions": "",
