+++ frontend/src/i18n/locales/ja/translation.json
@@ -59,11 +59,12 @@
   "Settings": "設定",
   "Delete": "削除",
   "Delete Cluster": "クラスターの削除",
+  "Control plane unhealthy": "",
+  "Unknown": "不明",
   "Active": "アクティブ",
   "Authentication required": "",
   "Insufficient permissions": "",
   "Unavailable": "",
-  "Control plane unhealthy": "",
   "Plugin": "プラグイン",
   "In-cluster": "クラスター内",
   "Cluster Inventory": "",
@@ -96,7 +97,6 @@
   "Show notifications": "通知を表示",
   "You have unread notifications": "未読の通知があります",
   "View all notifications": "すべての通知を表示",
-  "Unknown": "不明",
   "Are you sure you want to delete this plugin?": "このプラグインを削除してもよろしいですか？",
   "Plugins": "プラグイン",
   "Multiple versions": "",
