+++ frontend/src/i18n/locales/ko/translation.json
@@ -59,11 +59,12 @@
   "Settings": "설정",
   "Delete": "삭제",
   "Delete Cluster": "클러스터 삭제",
+  "Control plane unhealthy": "",
+  "Unknown": "알 수 없음",
   "Active": "활성",
   "Authentication required": "",
   "Insufficient permissions": "",
   "Unavailable": "",
-  "Control plane unhealthy": "",
   "Plugin": "플러그인",
   "In-cluster": "클러스터 내부",
   "Cluster Inventory": "",
@@ -96,7 +97,6 @@
   "Show notifications": "알림 보기",
   "You have unread notifications": "읽지 않은 알림이 있습니다.",
   "View all notifications": "모든 알림 보기",
-  "Unknown": "알 수 없음",
   "Are you sure you want to delete this plugin?": "이 플러그인을 삭제하시겠습니까?",
   "Plugins": "플러그인",
   "Multiple versions": "",
