+++ frontend/src/i18n/locales/pt/translation.json
@@ -59,11 +59,12 @@
   "Settings": "Definições",
   "Delete": "Eliminar",
   "Delete Cluster": "Eliminar Cluster",
+  "Control plane unhealthy": "",
+  "Unknown": "Desconhecido",
   "Active": "Activo",
   "Authentication required": "",
   "Insufficient permissions": "",
   "Unavailable": "",
-  "Control plane unhealthy": "",
   "Plugin": "Plugin",
   "In-cluster": "No cluster",
   "Cluster Inventory": "",
@@ -96,7 +97,6 @@
   "Show notifications": "Mostrar notificações",
   "You have unread notifications": "Tem notificações não lidas",
   "View all notifications": "Ver todas as notificações",
-  "Unknown": "Desconhecido",
   "Are you sure you want to delete this plugin?": "Tem a certeza que deseja eliminar este plugin?",
   "Plugins": "Plugins",
   "Multiple versions": "",
