+++ frontend/src/i18n/locales/ru/translation.json
@@ -60,6 +60,7 @@
   "Delete": "Удалить",
   "Delete Cluster": "Удалить кластер",
   "Control plane unhealthy": "",
+  "Unknown": "Неизвестный",
   "Active": "Активный",
   "Authentication required": "",
   "Insufficient permissions": "",
@@ -96,7 +97,6 @@
   "Show notifications": "Показать уведомления",
   "You have unread notifications": "У вас есть непрочитанные уведомления",
   "View all notifications": "Просмотреть все уведомления",
-  "Unknown": "Неизвестный",
   "Are you sure you want to delete this plugin?": "Вы уверены, что хотите удалить этот плагин?",
   "Plugins": "Плагины",
   "Multiple versions": "Несколько версий",
