+++ frontend/src/i18n/locales/ta/translation.json
@@ -59,11 +59,12 @@
   "Settings": "அமைப்புகள்",
   "Delete": "நீக்கு",
   "Delete Cluster": "க்ளஸ்டர் நீக்கு",
+  "Control plane unhealthy": "",
+  "Unknown": "தெரியாதது",
   "Active": "செயல்பாட்டில்",
   "Authentication required": "",
   "Insufficient permissions": "",
   "Unavailable": "",
-  "Control plane unhealthy": "",
   "Plugin": "பிளக்-இன்",
   "In-cluster": "க்ளஸ்டருக்குள்",
   "Cluster Inventory": "",
@@ -96,7 +97,6 @@
   "Show notifications": "அறிவிப்புகளைக் காட்டு",
   "You have unread notifications": "உங்களிடம் படிக்காத அறிவிப்புகள் உள்ளன",
   "View all notifications": "அனைத்து அறிவிப்புகளையும் பார்",
-  "Unknown": "தெரியாதது",
   "Are you sure you want to delete this plugin?": "இந்த பிளக்-இனை நீக்க விரும்புகிறீர்களா?",
   "Plugins": "பிளக்-இன்கள்",
   "Multiple versions": "",
