+++ frontend/src/i18n/locales/ur/translation.json
@@ -59,12 +59,15 @@
   "Settings": "ترتیبات",
   "Delete": "حذف کریں",
   "Delete Cluster": "کلسٹر حذف کریں",
+  "Control plane unhealthy": "",
+  "Unknown": "نامعلوم",
   "Active": "فعال",
   "Authentication required": "",
   "Insufficient permissions": "",
   "Unavailable": "",
   "Plugin": "پلگ ان",
   "In-cluster": "کلسٹر کے اندر",
+  "Cluster Inventory": "",
   "View Clusters": "کلسٹرز دیکھیں",
   "Loading...": "لوڈ ہو رہا ہے...",
   "No clusters found": "کوئی کلسٹر نہیں ملا",
@@ -94,7 +97,6 @@
   "Show notifications": "اطلاعات دکھائیں",
   "You have unread notifications": "آپ کے پاس غیر پڑھی اطلاعات موجود ہیں",
   "View all notifications": "تمام اطلاعات دیکھیں",
-  "Unknown": "نامعلوم",
   "Are you sure you want to delete this plugin?": "کیا آپ واقعی اس پلگ ان کو حذف کرنا چاہتے ہیں؟",
   "Plugins": "پلگ انز",
   "Multiple versions": "متعدد ورژنز",
