+++ frontend/src/i18n/locales/zh-tw/translation.json
@@ -59,11 +59,12 @@
   "Settings": "設定",
   "Delete": "刪除",
   "Delete Cluster": "刪除叢集",
+  "Control plane unhealthy": "",
+  "Unknown": "未知",
   "Active": "活躍",
   "Authentication required": "",
   "Insufficient permissions": "",
   "Unavailable": "",
-  "Control plane unhealthy": "",
   "Plugin": "外掛",
   "In-cluster": "叢集內",
   "Cluster Inventory": "",
@@ -96,7 +97,6 @@
   "Show notifications": "顯示通知",
   "You have unread notifications": "您有未讀通知",
   "View all notifications": "檢視所有通知",
-  "Unknown": "未知",
   "Are you sure you want to delete this plugin?": "您確定要刪除此外掛嗎？",
   "Plugins": "外掛",
   "Multiple versions": "多個版本",
