+++ frontend/src/i18n/locales/zh/translation.json
@@ -59,11 +59,12 @@
   "Settings": "设置",
   "Delete": "删除",
   "Delete Cluster": "删除集群",
+  "Control plane unhealthy": "",
+  "Unknown": "未知",
   "Active": "活跃",
   "Authentication required": "",
   "Insufficient permissions": "",
   "Unavailable": "",
-  "Control plane unhealthy": "",
   "Plugin": "插件",
   "In-cluster": "集群内",
   "Cluster Inventory": "",
@@ -96,7 +97,6 @@
   "Show notifications": "显示通知",
   "You have unread notifications": "您有未读通知",
   "View all notifications": "查看所有通知",
-  "Unknown": "未知",
   "Are you sure you want to delete this plugin?": "您确定要删除此插件吗？",
   "Plugins": "插件",
   "Multiple versions": "多个版本",
